--1.0.0版本，完成excel导出成照片

+++ src/hut/reoger/utils/ReadExcel.java
@@ -54,7 +54,8 @@ public class ReadExcel {
                         cellValue = cell.getDateCellValue();
                     }else{
                         //数字
-                        cellValue = String.valueOf(cell.getNumericCellValue());
+                        cellValue = String.valueOf((int)cell.getNumericCellValue());
+
                     }
                     break;
                 }
